+++ snowddl/resolver/user_role.py
@@ -1,10 +1,32 @@
-from snowddl.blueprint import RoleBlueprint, UserBlueprint, Grant, build_role_ident
+from snowddl.blueprint import RoleBlueprint, UserBlueprint, Grant, build_role_ident, build_grant_name_ident_snowflake
 from snowddl.resolver.abc_role_resolver import AbstractRoleResolver, ObjectType
 
 
 class UserRoleResolver(AbstractRoleResolver):
     def get_role_suffix(self):
         return self.config.USER_ROLE_SUFFIX
+
+    def get_existing_role_grants(self, role_name):
+        grants = []
+
+        cur = self.engine.execute_meta("SHOW GRANTS TO ROLE {role_name:i}", {
+            "role_name": role_name,
+        })
+
+        for r in cur:
+            # Check ROLE grants only, ignore everything else
+            # User roles may accumulate random grants from temporary tables and stages
+            # as well as ownership of manually created objects
+            if r['granted_on'] != 'ROLE':
+                continue
+
+            grants.append(Grant(
+                privilege=r['privilege'],
+                on=ObjectType[r['granted_on']],
+                name=build_grant_name_ident_snowflake(r['name'], ObjectType[r['granted_on']]),
+            ))
+
+        return role_name, grants, []
 
     def get_blueprints(self):
         blueprints = []